Moves functionality to 'EventsSubscriber' + renames
